+++ src/main/java/com/mallang/mallang_backend/domain/keyword/entity/Keyword.java
@@ -1,6 +1,6 @@
 package com.mallang.mallang_backend.domain.keyword.entity;
 
-import com.mallang.mallang_backend.domain.video.entity.Video;
+import com.mallang.mallang_backend.domain.video.video.entity.Video;
 import jakarta.persistence.*;
 import lombok.*;
