+++ src/main/java/com/mallang/mallang_backend/domain/video/video/controller/VideoController.java
@@ -1,13 +1,13 @@
-package com.mallang.mallang_backend.domain.video.controller;
+package com.mallang.mallang_backend.domain.video.video.controller;
 
-import com.mallang.mallang_backend.domain.video.dto.VideoResponse;
-import com.mallang.mallang_backend.domain.video.service.VideoService;
+import com.mallang.mallang_backend.domain.video.video.dto.VideoResponse;
+import com.mallang.mallang_backend.domain.video.video.service.VideoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import static com.mallang.mallang_backend.global.constants.AppConstants.*;
 import static com.mallang.mallang_backend.global.exception.ErrorCode.*;
 import java.io.IOException;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
