+++ src/main/java/com/mallang/mallang_backend/domain/video/video/repository/VideoRepository.java
@@ -1,6 +1,6 @@
-package com.mallang.mallang_backend.domain.video.repository;
+package com.mallang.mallang_backend.domain.video.video.repository;
 
-import com.mallang.mallang_backend.domain.video.entity.Video;
+import com.mallang.mallang_backend.domain.video.video.entity.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VideoRepository extends JpaRepository<Video, Long> {
