+++ src/main/java/com/mallang/mallang_backend/domain/video/video/service/impl/SearchContext.java
@@ -1,4 +1,4 @@
-package com.mallang.mallang_backend.domain.video.service.impl;
+package com.mallang.mallang_backend.domain.video.video.service.impl;
 
 import lombok.Value;
